minor change to prepare stage2 dataset script

+++ utils/prepare_stage2_dataset.py
@@ -56,7 +56,7 @@ def prepare_aloha_dataset(raw_dir, use_subset=False, target_ratio=0.70):
     print(f"[ALOHA] Views: {views}")
 
     # Assuming total target frames ~100k, ALOHA should contribute ~70k frames
-    target_frames = (100000 if not use_subset else 10000) * target_ratio
+    target_frames = int((100000 if not use_subset else 10000) * target_ratio)
     print(
         f"[ALOHA] Processing episodes for {target_ratio*100}%"
         f" of mixture = {target_frames} frames..."
@@ -145,7 +145,7 @@ def prepare_trex_dataset(raw_dir, use_subset=False, target_ratio=0.30):
     # Total size: 1.53 TB
 
     # Assuming total target frames ~100k, T-REX should contribute ~30k frames
-    target_frames = (100000 if not use_subset else 10000) * target_ratio
+    target_frames = int((100000 if not use_subset else 10000) * target_ratio)
     print(
         f"[T-REX] Processing episodes for {target_ratio*100}%"
         f" of mixture = {target_frames} frames..."
